Updated Auto Ownership Estimation
Renamed the module EstimateMississaugaAO to EstimateAutoOwnership and generalized it to take in an given number of vehicles as the maximum.

Improved the performance of the estimation model system to use SIMD routines for computing logorithms.
