+++ Tasha/Tasha.csproj
@@ -78,8 +78,8 @@
 		<ProjectReference Include="..\TashaInterfaces\TashaInterfaces.csproj" />
 	</ItemGroup>
 	<ItemGroup>
-		<PackageReference Include="System.Configuration.ConfigurationManager" Version="9.0.0" />
+		<PackageReference Include="System.Configuration.ConfigurationManager" Version="9.0.7" />
 		<PackageReference Include="System.Windows.Forms.DataVisualization" Version="1.0.0-prerelease.20110.1" />
-		<PackageReference Include="Microsoft.Windows.Compatibility" Version="9.0.0" />
+		<PackageReference Include="Microsoft.Windows.Compatibility" Version="9.0.7" />
 	</ItemGroup>
 </Project>
